fix: Skip loading statusline on floating windows.
Until now the satusline would load even when entering floating windows,
despite the fact that floating windows can't have statuslines. This
would also cause issues if the target was a window-relative float. In
this case the update would be applied to the parent window, but with the
wrong file info.